+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - allow empty account name on msg.Validate()
 - remove certificate indexing
+- fix account transfer
 
 ## 0.2.3